Use different strategies for index (real/mock).
